refactor: move all tests next to source files
