+++ lib/XwingListLoader.test.js
@@ -1,4 +1,4 @@
-import XwingListLoader from '../lib/XwingListLoader';
+import XwingListLoader from './XwingListLoader';
 
 const XWS_URL = 'http://mybuilder.com/path/to/squad.xws';
 let mockIntegration;
