+++ lib/integrations/MetaWing.test.js
@@ -1,4 +1,4 @@
-import MetaWing from '../../lib/integrations/MetaWing';
+import MetaWing from './MetaWing';
 
 describe('#matches() matches all MetaWing urls', () => {
   test('/squadrons/<id> url', () => {
