+++ lib/integrations/MetaWing2.test.js
@@ -1,4 +1,4 @@
-import MetaWing2 from '../../lib/integrations/MetaWing2';
+import MetaWing2 from './MetaWing2';
 
 describe('#matches() matches all MetaWing2 urls', () => {
   test('/squadrons/<id> url', () => {
