+++ lib/integrations/YASB.test.js
@@ -1,4 +1,4 @@
-import YASB from '../../lib/integrations/YASB';
+import YASB from './YASB';
 
 describe('#matches() matches all YASB urls', () => {
   test('https url', () => {
