+++ lib/integrations/YASB2.test.js
@@ -1,4 +1,4 @@
-import YASB2 from '../../lib/integrations/YASB2';
+import YASB2 from './YASB2';
 
 describe('#matches() matches all YASB urls', () => {
   test('https url', () => {
