+++ tsconfig.json
@@ -67,5 +67,5 @@
     "forceConsistentCasingInFileNames": true /* Disallow inconsistently-cased references to the same file. */
   },
   "include": ["lib"],
-  "exclude": ["node_modules", "dist"]
+  "exclude": ["node_modules", "dist", "lib/**/*.test.*"]
 }
